First Draft - Face Index Issue
Was contacted by spaztron64 on Jan4/2023 to look at SC1(Arcade). They provided samples and information through a blog post they made:
https://lainnet.superglobalmegacorp.com/blog/2022_12_28_n01.html
The post shared observations of the file format and geometry manipulations made in game by changing values in memory.

I never played Arcade or Playstation games in my youth so naturally never looked at such formats in the past. That being said my understanding of these PS1 formats is extremely limited. 

Here is a block out of the various structures within the model file format

[HEADER]
	|
	+-->[TABLE1]------------+    :: Strip Table
	|     	 |                          	|
	|      	+-->[TABLE2]     	|    :: Vertex Data
	|                                  	|
	+-->[TABLE3]             	|    :: Geometry Data
	|                                  	|
	+-->[TABLE4]<----------+    :: Bone Index / Name

Strip Table, works more like a vertex group in blender, therefore each entry is actually a bone transform accompanied by a pointer that goes to table2 (Vertex Data) containing vertices which that bone affects. 
I initially named it the strip table because in addition the vertices are interweaved or presented in strips. I'm still not sure why they are doing this, I would have to imagine it's for each sub material. 
Vertex Data in table2 are in local space so need to be transformed into world space using the transform provided in the entries of table1.

Table3 contain entries for each face, and a bunch of junk that is important for the PlayStation grapics renderer. 
It annoyingly presents alot of packed commands which are essential for decoding the face indices. Table4 is a list of names which table1 points to, seems not important for the model

Faces don't currently form correctly, a decode routine was provided by Spicyjpeg which I believe is working. 
However it seems I'm missing a flag that tells me when to build a 3 sided or 4 sided polygon.
This lack of functionality results in the model having incorrect face indices, which manifests as missing or extra faces. 
Thus to reduce complexity of the current state of the maxscript certain import features are not presently implemented:

- [NO] Normals
- [NO] UV's (Texture Coodinates)
- [NO] Weights